Add integration-specific data cleanup on disconnect
ISSUE: UPP-476

+++ src/BusinessLogic/Domain/Disconnect/Services/DisconnectService.php
@@ -4,12 +4,12 @@ namespace Unzer\Core\BusinessLogic\Domain\Disconnect\Services;
 
 use Unzer\Core\BusinessLogic\Domain\Connection\Exceptions\ConnectionSettingsNotFoundException;
 use Unzer\Core\BusinessLogic\Domain\Connection\Services\ConnectionService;
+use Unzer\Core\BusinessLogic\Domain\Integration\Disconnect\DisconnectIntegrationServiceInterface;
 use Unzer\Core\BusinessLogic\Domain\PaymentMethod\Interfaces\PaymentMethodConfigRepositoryInterface;
 use Unzer\Core\BusinessLogic\Domain\PaymentPageSettings\Repositories\PaymentPageSettingsRepositoryInterface;
 use Unzer\Core\BusinessLogic\Domain\PaymentStatusMap\Interfaces\PaymentStatusMapRepositoryInterface;
 use Unzer\Core\BusinessLogic\Domain\TransactionHistory\Interfaces\TransactionHistoryRepositoryInterface;
 use Unzer\Core\Infrastructure\ORM\Exceptions\QueryFilterInvalidParamException;
-use UnzerSDK\Exceptions\UnzerApiException;
 
 /**
  * Class DisconnectService.
@@ -44,24 +44,32 @@ class DisconnectService
     private TransactionHistoryRepositoryInterface $transactionHistoryRepository;
 
     /**
+     * @var DisconnectIntegrationServiceInterface
+     */
+    private DisconnectIntegrationServiceInterface $disconnectIntegrationService;
+
+    /**
      * @param ConnectionService $connectionService
      * @param PaymentMethodConfigRepositoryInterface $paymentMethodConfigRepository
      * @param PaymentPageSettingsRepositoryInterface $paymentPageSettingsRepository
      * @param PaymentStatusMapRepositoryInterface $paymentStatusMapRepository
      * @param TransactionHistoryRepositoryInterface $transactionHistoryRepository
+     * @param DisconnectIntegrationServiceInterface $integrationService
      */
     public function __construct(
         ConnectionService $connectionService,
         PaymentMethodConfigRepositoryInterface $paymentMethodConfigRepository,
         PaymentPageSettingsRepositoryInterface $paymentPageSettingsRepository,
         PaymentStatusMapRepositoryInterface $paymentStatusMapRepository,
-        TransactionHistoryRepositoryInterface $transactionHistoryRepository
+        TransactionHistoryRepositoryInterface $transactionHistoryRepository,
+        DisconnectIntegrationServiceInterface $disconnectIntegrationService
     ) {
         $this->connectionService = $connectionService;
         $this->paymentMethodConfigRepository = $paymentMethodConfigRepository;
         $this->paymentPageSettingsRepository = $paymentPageSettingsRepository;
         $this->paymentStatusMapRepository = $paymentStatusMapRepository;
         $this->transactionHistoryRepository = $transactionHistoryRepository;
+        $this->disconnectIntegrationService = $disconnectIntegrationService;
     }
 
     /**
@@ -76,6 +84,8 @@ class DisconnectService
 
         $this->connectionService->deleteConnectionSettings();
 
+        $this->disconnectIntegrationService->deleteIntegrationData();
+
         $this->deleteAdditionalSettings();
     }
 
@@ -86,7 +96,10 @@ class DisconnectService
 
     public function deleteAdditionalSettings(): void
     {
-        $this->paymentPageSettingsRepository->deletePaymentPageSettings();
+        if ($this->disconnectIntegrationService->shouldDeletePayPageSettings()) {
+            $this->paymentPageSettingsRepository->deletePaymentPageSettings();
+        }
+
         $this->paymentStatusMapRepository->deletePaymentStatusMapEntity();
         $this->paymentMethodConfigRepository->deletePaymentConfigEntities();
         $this->transactionHistoryRepository->deleteTransactionHistoryEntities();
